rework xxHash for longs, rework getBitsPerItemForFpRate to remove Guava code
Chaning getBitsPerItemForFpRate may have corrected a intermittent bug, with MGunlogson/CuckooFilter4J and xxhashVersion (with generics) false positives would exist for some runs but would be 0 for others.  

for Example with a false positive rate of 0.001 a max items size of 20000000 and 32751180 mightContain() querys I would get between 102000 & 103000 false positives on some runs but 0 on others with same code and data. However when reworking getBitsPerItemForFpRate to use java Math.ceil and cast to int I always get 0 false positives. Maybe the Guava code was buggy and allocating to few bits for fingerprint?